Allow assigning our own reference to Notify emails
https://mhclgdigital.atlassian.net/browse/FSPT-137

Our end-to-end test suite is currently taking quite a long time to run,
as many of the tests run sequentially. The main reason for needing them
to run sequentially is that some tests login as the same user, and the
login attempts are in conflict. This conflict is not handled well by the
end-to-end tests and concurrent logins are not well supported by
authenticator. Authenticator allows multiple login sessions to exist at
once, but does not allow multiple unredeemed magic links to exist for
one account at the same time. Furthermore, the end-to-end tests use a
"list all magic links" endpoint on authenticator and then grab the top
magic link off the stack. This link isn't necessarily associated with
the account that they are trying to log in as, which is bad.

This patch allows a reference to be associated with an email sent via
GOV.UK Notify. When a magic link email is sent, the calling code can set
a reference which the e2e tests can then scan for. This will let the e2e
tests retrieve the _correct_ magic link from Notify for the account it
was trying to log into.